requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath==0.82
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pytest==7.1.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
